Dockerize deployment of hamr
New flow:
After syncing an app (maybe from the UI), the requirements.txt
needs to be synced. This can be done with ./catreq.sh if the installed
app also has a requirements.txt in its root directory.
This will concatenate that file with Hamr's requirements.txt (hamr-requirements.txt)
requirements.txt.
After this the docker container needs to be restarted. This is because
the docker command has a pip install that will install all dependencies
from requirements.txt before starting apache.